feat: make web responsive
