+++ src/App.css
@@ -52,10 +52,7 @@
   border-radius: 30px;
 }
 
-.like-page-container {
-  display: grid;
-  grid-template-columns: 1fr 2fr;
-}
+
