+++ src/components/PendingRecipeDetail/PendingRecipeDetail.css
@@ -13,3 +13,12 @@
     margin-bottom: 10px;
     z-index: 1;
 }
+
+@media (max-width: 900px) {
+    .recipeDetailPending{
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+    }
+}
+
