+++ src/components/RecipeCard/RecipeCard.css
@@ -1,9 +1,12 @@
+:root{
+  --scale: 1;
+}
 
 .recipeCard {
   height: fit-content;
-  max-width: 260px;
-  margin: 20px;
-  border-radius: 16px;
+  max-width: calc(var(--scale)*260px) ;
+  margin: calc(var(--scale)*20px);
+  border-radius:  16px
   
   /* border: 1px solid #000000; */
   /* border: 1px solid #A29E9E ; */
@@ -20,13 +23,14 @@
 }
 
 .recipeCard img {
-  max-width: 260px;
+  max-width: calc(var(--scale)*260px);
   max-height: 100%;
   border-radius: 10px;
 }
 
 .recipeCard p {
-  max-height: 77px;
+  max-height: calc(var(--scale)*54px);
+  font-size: 16px;
   /* overflow: hidden; */
 }
 
@@ -47,7 +51,7 @@
 
 .recipeCard__info p {
   max-height: 54px;
-  max-width: 210px;
+  max-width: calc(var(--scale)*260px);
   margin-top: 5px;
   margin-bottom: 5px;
   overflow: hidden;
@@ -130,4 +134,75 @@
 }
 
 
+/* ------Responsive----------- */
+
+@media (max-width: 1536px) {
+
+}
+
+/* / xl / */
+@media (max-width: 1280px) {
+
+}
+
+/* / lg / */
+@media (max-width: 1024px) {
+  :root{
+    --scale: 0.8;
+  }
+
+  .recipeCard__info p {
+   font-size: small;
+  }
+}
+
+@media(max-width: 843px){
+  :root{
+    --scale: 1;
+  }
+
+  .recipeCard__info p {
+   font-size: normal;
+  }
+}
+
+@media(max-width: 717px){
+  :root{
+    --scale: 1.5;
+  }
+
+  .recipeCard__info p {
+   font-size: normal;
+  }
+}
+
+/* / md / */
+@media (max-width: 768px) {
+
+}
+
+/* / sm / */
+@media (max-width: 640px) {
+
+}
+
+@media (max-width: 516px) {
+  :root{
+    --scale: 1.2;
+  }
+}
+
+/* / xs */ 
+ @media (max-width: 475px) {
+  :root{
+    --scale: 1;
+  }
+}
+
+  @media (max-width: 376) {
+    :root{
+      --scale: 0.6;
+    }
+  
+}
