+++ src/components/RecipeCardList/RecipeCardList.jsx
@@ -2,9 +2,9 @@ import "./RecipeCardList.css";
 
 import RecipeCard from "../RecipeCard/RecipeCard";
 
-const RecipeCardList = ({props,style,pending}) => {
+const RecipeCardList = ({props,pending}) => {
   return (
-    <div className="recipe-list" style={style && { flexWrap: 'nowrap' }} pending={pending}>
+    <div className="recipe-list" pending={pending}>
       {props.map((item) => (   
         <RecipeCard
           key={item.id}
